requirements.txt
Flask==2.3.3
opencv-python==4.8.1.78
torch==2.0.1
torchvision==0.15.2
numpy==1.24.3
Pillow==10.0.0